Vuefy and cleanup OptionsGameplayTab
