+++ src/components/tabs/index.js
@@ -1,4 +1,5 @@
 // Temporarily aggregate tab components here until GameUI is migrated to SFC
+import OptionsGameplayTab from "./options-gameplay/OptionsGameplayTab";
 import OptionsSavingTab from "./options-saving/OptionsSavingTab";
 import OptionsVisualTab from "./options-visual/OptionsVisualTab";
 import StatisticsTab from "./statistics/StatisticsTab";
@@ -11,6 +12,7 @@ import SecretAchievementTab from "./secret-achievements/SecretAchievementTab";
 const TabComponents = {
   OptionsSavingTab,
   OptionsVisualTab,
+  OptionsGameplayTab,
   StatisticsTab,
   ChallengeRecordsTab,
   PastPrestigeRunsTab,
